requirements.txt
boto3==1.35.39
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.0.3
PyYAML==6.0
minio==7.2.9
#python-dotenv==1.0.1
